Scripts/MawOfSouls SQL + Cleanup
